+++ src/util/test/test_porter_stemming.cpp
@@ -94,7 +94,7 @@ BOOST_AUTO_TEST_CASE(TestPorterStemming)
              sw.Stop();
              double e = sw.Elapsed();
              cerr << "  ...done: " << count << " items in " << e
-                 << " seconds = " << count / e << " items/sec" << endl;
+                  << " seconds = " << double(count) / e << " items/sec" << endl;
          }}
 
         {{
@@ -113,7 +113,7 @@ BOOST_AUTO_TEST_CASE(TestPorterStemming)
              sw.Stop();
              double e = sw.Elapsed();
              cerr << "  ...done: " << count << " items in " << e
-                 << " seconds = " << count / e << " items/sec" << endl;
+                  << " seconds = " << (double)count / e << " items/sec" << endl;
          }}
 
     } else {
